Extract search and filter views to own files
